+++ src/main/java/com/thadocizn/todo/models/Users.java
@@ -8,7 +8,7 @@ public class Users {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long userid;
+    private int userid;
 
     @Column(nullable = false)
     private String username;
@@ -16,11 +16,11 @@ public class Users {
     public Users() {
     }
 
-    public long getUserid() {
+    public int getUserid() {
         return userid;
     }
 
-    public void setUserid(long userid) {
+    public void setUserid(int userid) {
         this.userid = userid;
     }
